Revert "Feat: Add new home page by Team or User (#105)"
This reverts commit 3a99cf4a6da3435dd3cbb04565637d39d3edbac5.
