--- cypress/fixtures/teams/all-teams.json
@@ -1,16 +1,0 @@
-[
-	{
-		"name": "Team",
-		"id": "e2308239-b038-4abc-b288-dfae2bb7e008",
-		"userMembers": [
-			{
-				"id": "ced8c2ec-db6b-4ee2-9cb8-c375cd52e30e",
-				"teamId": "e2308239-b038-4abc-b288-dfae2bb7e008",
-				"userId": "ffe04ac2-e661-426d-abf9-937bf9db6382",
-				"role": "OWNER"
-			}
-		],
-		"invitations": [],
-		"collections": []
-	}
-]
