+++ cypress/tests/ui/exceptions/constants.ts
@@ -21,11 +21,6 @@ export const keypair = {
 	},
 };
 
-export const homePage = {
-	title: 'Workspaces',
-	list: ['My Personal account', 'Team'],
-};
-
 export const collections = {
 	pageTitle: 'Collections',
 	btnCreateText: 'Create a new collection',
