+++ package.json
@@ -51,7 +51,7 @@
 		"flagsmith": "^3.22.1",
 		"localforage": "^1.10.0",
 		"lodash.debounce": "^4.0.8",
-		"lucide-react": "^0.371.0",
+		"lucide-react": "^0.284.0",
 		"match-sorter": "^6.3.1",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
