+++ src/common/api/collections.ts
@@ -37,7 +37,7 @@ export const useNewCollectionMutation = () => {
 			axios?.post('/collection', { name }).then((res) => res.data),
 		onSuccess: (data) => {
 			queryClient.invalidateQueries({ queryKey: ['collections'] });
-			navigate(`/user/collection/${data.id}`);
+			navigate(`/collection/${data.id}`);
 		},
 	});
