--- src/common/types/team.ts
@@ -1,10 +1,0 @@
-import { Collection } from './collection';
-import { User } from './user';
-
-export type Team = {
-	id: string;
-	name: string;
-	adminId: string;
-	users: User[];
-	collections: Collection[];
-};
