--- src/common/types/user.ts
@@ -1,3 +1,0 @@
-export type User = {
-	email: string;
-};
