+++ src/configs/router/index.tsx
@@ -7,7 +7,6 @@ import Login from '@/common/components/auth/Login';
 import RecoverPassword from '@/common/components/auth/RecoverPassword';
 import ResetPassword from '@/common/components/auth/ResetPassword';
 import CollectionCTAPage from '@/pages/Collection/CollectionCTAPage';
-import CollectionHome from '@/pages/Collection/CollectionHome';
 import CollectionPage from '@/pages/Collection/CollectionPage';
 import InvocationPage from '@/pages/Invocation/InvocationPage';
 import AuthPage from '@/pages/auth/AuthPage';
@@ -34,32 +33,22 @@ const router = createBrowserRouter([
 				element: <ChangePassword />,
 			},
 			{
-				path: 'user',
-				element: <CollectionHome />,
+				path: 'collection/:collectionId',
+				element: <CollectionPage />,
 				children: [
 					{
-						path: 'collection/:collectionId',
-						element: <CollectionPage />,
-						children: [
-							{
-								index: true,
-								element: <CollectionCTAPage />,
-							},
-							{
-								path: 'variables',
-								element: <CollectionVariablesContainer />,
-							},
-							{
-								path: 'invocation/:invocationId',
-								element: <InvocationPage />,
-							},
-						],
+						index: true,
+						element: <CollectionCTAPage />,
+					},
+					{
+						path: 'variables',
+						element: <CollectionVariablesContainer />,
+					},
+					{
+						path: 'invocation/:invocationId',
+						element: <InvocationPage />,
 					},
 				],
-			},
-			{
-				path: 'team/:teamId',
-				element: <CollectionHome />,
 			},
 		],
 	},
